+++ package.json
@@ -15,6 +15,9 @@
     "dev2": "while true; do yarn --check-files && yarn dev; done",
     "test-types": "yarn tsc",
     "test-unit": "yarn jest test/unit/",
+    "test-dev": "cross-env NEXT_TEST_MODE=dev yarn testheadless",
+    "test-start": "cross-env NEXT_TEST_MODE=start yarn testheadless",
+    "test-deploy": "cross-env NEXT_TEST_MODE=deploy yarn testheadless",
     "testonly": "yarn jest --runInBand",
     "testheadless": "cross-env HEADLESS=true yarn testonly",
     "genstats": "cross-env LOCAL_STATS=true node .github/actions/next-stats-action/src/index.js",
