+++ packages/next/server/web-server.ts
@@ -94,7 +94,7 @@ export default class NextWebServer extends BaseServer<WebServerOptions> {
       [this.serverOptions.webServerConfig.page]: '',
     }
   }
-  protected getappPathsManifest() {
+  protected getAppPathsManifest() {
     return {
       [this.serverOptions.webServerConfig.page]: '',
     }
